Update README.md
I. Motivation

This project will give you experience in applying dynamic memory management, implementing a template container class (the double‐ended, doubly‐linked list, or Dlist), and using the at–most‐ once invariant and existence, ownership, and conservation rules to implement two simple applications using this structure.

II. Programming Assignment

You will first implement a templated double-ended, doubly-linked list, or Dlist. Then, you will use Dlist to build two applications: a reverse-Polish notation calculator and a call center simulation program.

1. The Double-Ended, Doubly-Linked List

The double‐ended, doubly-linked list, or Dlist, is a templated container. It supports the following operational methods:

isEmpty: a predicate that returns true if the list is empty, false otherwise.

insertFront/insertBack: insert an object at the front/back of the list, respectively.

removeFront/removeBack: remove an object from the front/back of a non-empty list, respectively; throws an exception if the list is empty.

Note that while this list is templated across the contained type, T, it is a container of pointers-to-T, not container of instances of T. Insertion takes a pointer-to‐T as an argument and put that

pointer into the node to be inserted. Removal removes a node and returns the pointer-to-T kept in that node. This ensures that the Dlist implementation knows that: it owns inserted objects, it is responsible for copying them if the list is copied, and it must destroy them if the list is destroyed.

The complete interface of the Dlist class is provided in the dlist.h, which is available in the Project-5-Related-Files.zip on Canvas. The code is replicated here for your convenience.